+++ src/gfx/display.cxx
@@ -3,10 +3,10 @@
 void Display::text(const Text& text) {
     const auto& ops = text.ops();
 
-    for (i32 op_index = 0; op_index < ops.size(); ++op_index) {
+    for (size_t op_index = 0; op_index < ops.size(); ++op_index) {
         const auto& op = ops[op_index];
         const auto glyph_count = op.glyphs.size();
-        for (i32 i = 0; i < glyph_count; ++i) {
+        for (size_t i = 0; i < glyph_count; ++i) {
             const i32 glyph = op.glyphs[i];
             const Vec2<u32> pos = op.position + Vec2<i32>(i, 0);
             put(DisplayCell{ glyph, op.color }, pos);
